replace DATATYPE.INTEGER by FLOAT because float are needed

+++ models/advance.js
@@ -3,7 +3,7 @@ module.exports = (sequelize, DataTypes) => {
   const Advance = sequelize.define('Advance', {
     description: DataTypes.STRING,
     date: DataTypes.DATEONLY,
-    cost: DataTypes.INTEGER,
+    cost: DataTypes.FLOAT,
     buyer: DataTypes.ENUM('Nils', 'Vio'),
   }, {});
   Advance.associate = function (models) {
