+++ models/expense.js
@@ -3,7 +3,7 @@ module.exports = (sequelize, DataTypes) => {
   const Expense = sequelize.define('Expense', {
     description: DataTypes.STRING,
     date: DataTypes.DATEONLY,
-    cost: DataTypes.INTEGER,
+    cost: DataTypes.FLOAT,
     buyer: DataTypes.ENUM('Nils', 'Vio'),
   }, {});
   Expense.associate = function(models) {
